feat(webui): 控制台重设计——自托管资产、状态词汇表与可访问性底线
实现 openspec change `redesign-webui-console`（28 项任务完成 26 项）。

设计系统
- style.css 重写为单一 token 层：七个核心 token + 六个状态色，
  浅/深色方案各自覆盖；色值全部经过对比度实测而非目测
- --rule（装饰性分隔线）与 --edge/--control（可操作元素边界）拆分：
  后者需满足 WCAG 1.4.11 的 3:1，此前输入框用 --rule 是真实缺陷
- --signal 只承担三种职责（焦点标记 / 主按钮 / 焦点环），永不作为正文色；
  .btn-warning 与 .alert-warning 改用 --danger——控制面断连是故障而非指引

状态词汇表
- 新增 SessionStatus 枚举，把 (MappingState, phase) 投影为显示状态。
  此前 phase 会把飞书 reaction 名（Get/OnIt/CrossMark）直接泄漏给操作者
- 单测覆盖 D7 全部七行输入；集成测试断言这些 token 不再出现在响应体中

自托管
- 三个可变字重 woff2（57 KB）+ marked / highlight.js / htmx 全部本地化
- 新增测试断言任何页面都不引用外部资产

模板
- 四个 admin 页面改为继承 base.html，导航与品牌区收敛到唯一来源
  （admin_login.html 有意保持独立：登录页不应渲染它守卫的导航）
- 删除死表面：/agent 导航项、未注册的 agent.html / agent_timeline.html
  （见 sebas-2ty）、以及随之失效的 .timeline-* / .message-* 等 18 条规则
- SSE 改为原生 EventSource：htmx sse 扩展从未加载，
  dashboard 上的 hx-trigger="sse:update" 一直是失效的

缺陷修复
- Close 按钮对所有会话都不可用：session key 经 percent 编码，
  hx-target="#row-oc_a%00" 是 CSS 选择器语法错误，htmx 解析即抛异常，
  请求根本发不出去；且该 target 指向 JSON 响应体。改为 hx-swap="none"，
  并补回归测试
- admin_status.html 里的 {{ status.operations.len() }} 是 Rust 方法语法，
  在 minijinja 中渲染为空
- 修复 admin.rs / admin_auth.rs 三处既有 clippy 警告

未完成（阻塞，见 sebas-1rf）
- 5.1 / 8.4 的 markdown 渲染：卡片正文是 agent 可控内容，
  而 marked 原样透传 HTML，故暂按转义文本渲染。副作用是
  marked + highlight.js 约 157 KB 每页空载，需决策后再收口

+++ webui/src/admin_auth.rs
@@ -165,7 +165,7 @@ pub fn generate_token(byte_len: usize) -> String {
         .duration_since(SystemTime::UNIX_EPOCH)
         .unwrap_or_default()
         .as_nanos();
-    let seed = (nanos as u128).to_le_bytes();
+    let seed = nanos.to_le_bytes();
     let mut state = 0u64;
     for b in seed.chunks(2) {
         state = state
